+++ docs/src/history.md
@@ -1,6 +1,12 @@
 # Version history
 
 
+## Current development version
+
+* ADDED: more method variations to lift unwrapped values
+* CHANGED: `unwrap()` for `Choice` renamed to `unwrap_choice()`, to not confuse it with the constant time `unwrap()` for `Value`.
+
+
 ## v0.0.1
 
 Initial version.